Variable set() and get() call through parent Device writeBlocks()/readBlocks()
Variable beforeReadCmd and afterWriteCmd have been removed, as this functionality can now be achieved by overriding the parent Device writeBlocks()/readBlocks().

Device writeBlocks()/readBlocks()/etc now take an optional 'variable' argument to reduce the scope of the call to a single Variable. We may want to give more thought to this.